kernelversion: Add utility for checking kernel version
This change moves functions for checking the kernel version out from the
daemon to the separate module.
